workspace/auto-detach-billing/function/requirements.txt
functions-framework>=3.4.0
google-api-python-client>=2.90.0
google-auth>=2.22.0